+++ tests/cli/clitest.cpp
@@ -32,11 +32,10 @@ static int s_errors;
 #define EXPECT(...) \
     if (!bool(__VA_ARGS__)) \
     failed(line ? line : __LINE__, #__VA_ARGS__)
+#define EXPECT_PARSE(cli, ...) parseTest(__LINE__, cli, __VA_ARGS__)
+#define EXPECT_ERR(cli, text) errTest(__LINE__, cli, text)
 #define EXPECT_HELP(cli, cmd, text) helpTest(__LINE__, cli, cmd, text)
 #define EXPECT_USAGE(cli, cmd, text) usageTest(__LINE__, cli, cmd, text)
-#define EXPECT_PARSE(cli, ...) parseTest(__LINE__, cli, true, 0, __VA_ARGS__)
-#define EXPECT_PARSE2(cli, cont, ec, ...) \
-    parseTest(__LINE__, cli, cont, ec, __VA_ARGS__)
 #define EXPECT_ARGV(fn, cmdline, ...) \
     toArgvTest(__LINE__, fn, cmdline, __VA_ARGS__)
 
@@ -44,6 +43,37 @@ static int s_errors;
 void failed(int line, char const msg[]) {
     cerr << "Line " << line << ": EXPECT(" << msg << ") failed" << endl;
     s_errors += 1;
+}
+
+//===========================================================================
+void parseTest(
+    int line,
+    Dim::Cli & cli,
+    string const & cmdline = {},
+    bool continueFlag = true,
+    int exitCode = -1
+) {
+    auto args = cli.toWindowsArgv(cmdline);
+    args.insert(args.begin(), kCommand);
+    bool rc = cli.parse(args);
+    if (exitCode == -1)
+        exitCode = rc ? Dim::kExitOk : Dim::kExitUsage;
+    if (rc != continueFlag || exitCode != cli.exitCode()) {
+        if (exitCode)
+            cerr << cli.errMsg() << endl;
+        EXPECT(rc == continueFlag);
+        EXPECT(exitCode == cli.exitCode());
+    }
+}
+
+//===========================================================================
+void errTest(int line, Dim::Cli & cli, string const & errText) {
+    ostringstream os;
+    cli.printError(os);
+    auto tmp = os.str();
+    EXPECT(tmp == errText);
+    if (tmp != errText)
+        cerr << tmp;
 }
 
 //===========================================================================
@@ -58,7 +88,7 @@ void helpTest(
     auto tmp = os.str();
     EXPECT(tmp == helpText);
     if (tmp != helpText)
-        cout << tmp;
+        cerr << tmp;
 }
 
 //===========================================================================
@@ -73,26 +103,7 @@ void usageTest(
     auto tmp = os.str();
     EXPECT(tmp == usageText);
     if (tmp != usageText)
-        cout << tmp;
-}
-
-//===========================================================================
-void parseTest(
-    int line,
-    Dim::Cli & cli,
-    bool continueFlag,
-    int exitCode,
-    vector<char const *> args
-) {
-    args.insert(args.begin(), kCommand);
-    args.push_back(nullptr);
-    bool rc = cli.parse(args.size() - 1, const_cast<char **>(args.data()));
-    if (rc != continueFlag || exitCode != cli.exitCode()) {
-        if (exitCode)
-            cerr << cli.errMsg() << endl;
-        EXPECT(rc == continueFlag);
-        EXPECT(exitCode == cli.exitCode());
-    }
+        cerr << tmp;
 }
 
 //===========================================================================
@@ -178,7 +189,7 @@ void parseTests() {
                 *arg *= tmp;
                 return true;
             });
-        EXPECT_PARSE(cli, {"-n2", "-n3"});
+        EXPECT_PARSE(cli, "-n2 -n3");
         EXPECT(*sum == 6);
     }
 
@@ -187,9 +198,9 @@ void parseTests() {
         cli = {};
         auto & opt = cli.opt("[value]", ExtractNoInsert::kBad)
             .desc("Value to attempt to parse.");
-        EXPECT_PARSE(cli, {"g"});
+        EXPECT_PARSE(cli, "g");
         EXPECT(*opt == ExtractNoInsert::kGood);
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"b"});
+        EXPECT_PARSE(cli, "b", false);
         EXPECT(*opt == ExtractNoInsert::kInvalid);
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS] [value]
@@ -205,9 +216,9 @@ Options:
         cli = {};
         auto & opt = cli.opt("[value]", ExtractWithInsert::kGood)
             .desc("Value to attempt to parse.");
-        EXPECT_PARSE(cli, {});
+        EXPECT_PARSE(cli);
         EXPECT(*opt == ExtractWithInsert::kGood);
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"b"});
+        EXPECT_PARSE(cli, "b", false);
         EXPECT(*opt == ExtractWithInsert::kInvalid);
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS] [value]
@@ -218,11 +229,11 @@ Options:
 )");
 
         opt.defaultValue(ExtractWithInsert::kInOnly);
-        EXPECT_PARSE(cli, {});
+        EXPECT_PARSE(cli);
         EXPECT(*opt == ExtractWithInsert::kInOnly);
-        EXPECT_PARSE(cli, {"g"});
+        EXPECT_PARSE(cli, "g");
         EXPECT(*opt == ExtractWithInsert::kGood);
-        EXPECT_PARSE(cli, {"i"});
+        EXPECT_PARSE(cli, "i");
         EXPECT(*opt == ExtractWithInsert::kInOnly);
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS] [value]
@@ -268,12 +279,14 @@ Options:
     EXPECT_USAGE(cli, "", 1 + R"(
 usage: test [--streetlight=COLOR] [--help]
 )");
-    EXPECT_PARSE(cli, {"--streetlight", "red"});
+    EXPECT_PARSE(cli, "--streetlight red");
     EXPECT(*state == State::stop);
 
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"--streetlight", "white"});
-    EXPECT(cli.errMsg() == "Invalid '--streetlight' value: white");
-    EXPECT(cli.errDetail() == R"(Must be "green", "yellow", or "red".)");
+    EXPECT_PARSE(cli, "--streetlight white", false);
+    EXPECT_ERR(cli, 1 + R"(
+Error: Invalid '--streetlight' value: white
+Must be "green", "yellow", or "red".
+)");
 
     state.defaultValue(State::go);
     EXPECT_HELP(cli, "", 1 + R"(
@@ -287,13 +300,14 @@ Options:
 
   --help               Show this message and exit.
 )");
+
     cli = {};
     auto & state2 = cli.optVec<State>("[streetlights]")
         .desc("Color of street lights.")
         .valueDesc("COLOR")
-        .choice(State::go, "green", "Means go!")
-        .choice(State::wait, "yellow", "Means wait, even if you're late.")
-        .choice(State::stop, "red", "Means stop.");
+        .choice(State::stop, "red", "Means stop.", "3")
+        .choice(State::wait, "yellow", "Means wait, even if you're late.", "2")
+        .choice(State::go, "green", "Means go!", "1");
     EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS] [streetlights...]
   streetlights  Color of street lights.
@@ -304,29 +318,39 @@ usage: test [OPTIONS] [streetlights...]
 Options:
   --help    Show this message and exit.
 )");
-    EXPECT_PARSE(cli, {"red"});
+    EXPECT_PARSE(cli, "red");
     EXPECT(state2.size() == 1 && state2[0] == State::stop);
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"white"});
-    EXPECT(cli.errMsg() == "Invalid 'streetlights' value: white");
-    EXPECT(cli.errDetail() == R"(Must be "green", "yellow", or "red".)");
+    EXPECT_PARSE(cli, "white", false);
+    EXPECT_ERR(cli, 1 + R"(
+Error: Invalid 'streetlights' value: white
+Must be "green", "yellow", or "red".
+)");
 
     EXPECT(state2.defaultValue() == State::go);
     state2.defaultValue(State::wait);
     EXPECT(state2.defaultValue() == State::wait);
 
     cli = {};
-    cli.optVec<unsigned>("n")
+    auto & nums = cli.optVec<unsigned>("n")
         .desc("List of numbers")
         .valueDesc("NUMBER")
-        .choice(1, "one").choice(2, "two").choice(3, "three")
-        .choice(4, "four").choice(5, "five").choice(6, "six")
-        .choice(7, "seven").choice(8, "eight").choice(9, "nine")
-        .choice(10, "ten").choice(11, "eleven").choice(12, "twelve");
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"-n", "white"});
-    EXPECT(cli.errMsg() == "Invalid '-n' value: white");
-    EXPECT(cli.errDetail() == 1 + R"(
+        .choice(1, "one").choice(2, "two");
+    EXPECT_PARSE(cli, "-n white", false);
+    EXPECT_ERR(cli, 1 + R"(
+Error: Invalid '-n' value: white
+Must be "one" or "two".
+)");
+
+    nums.choice(3, "three").choice(4, "four").choice(5, "five")
+        .choice(6, "six").choice(7, "seven").choice(8, "eight")
+        .choice(9, "nine").choice(10, "ten").choice(11, "eleven")
+        .choice(12, "twelve");
+    EXPECT_PARSE(cli, "-n white", false);
+    EXPECT_ERR(cli, 1 + R"(
+Error: Invalid '-n' value: white
 Must be "one", "two", "three", "four", "five", "six", "seven", "eight", "nine",
-"ten", "eleven", or "twelve".)");
+"ten", "eleven", or "twelve".
+)");
 }
 
 
@@ -350,7 +374,7 @@ void helpTests() {
         in.str("");
         out.str("");
         cli.iostreams(&in, &out);
-        EXPECT_PARSE2(cli, false, 0, {"--version"});
+        EXPECT_PARSE(cli, "--version", false, Dim::kExitOk);
         auto tmp = out.str();
         string expect = "test version 1.0\n";
         EXPECT(tmp == expect);
@@ -364,7 +388,7 @@ void helpTests() {
         cli.helpNoArgs();
         out.str("");
         cli.iostreams(nullptr, &out);
-        EXPECT_PARSE2(cli, false, Dim::kExitOk, {});
+        EXPECT_PARSE(cli, {}, false, Dim::kExitOk);
         cli.iostreams(nullptr, nullptr);
         EXPECT(out.str() == 1 + R"(
 usage: test [OPTIONS]
@@ -382,9 +406,9 @@ Options:
         bool help;
         cli.opt(&count, "c ?count").implicitValue(3);
         cli.opt(&help, "? h help");
-        EXPECT_PARSE(cli, {"-hc2", "-?"});
+        EXPECT_PARSE(cli, "-hc2 -?");
         EXPECT(count == 2);
-        EXPECT_PARSE(cli, {"--count"});
+        EXPECT_PARSE(cli, "--count");
         EXPECT(count == 3);
     }
 
@@ -418,7 +442,7 @@ Multiline footer:
 - first reference
 - second reference
 )");
-        EXPECT_PARSE(cli, {"--no-help"});
+        EXPECT_PARSE(cli, "--no-help");
     }
 
     // group sortKey
@@ -471,6 +495,20 @@ void cmdTests() {
         // create option and hide it underneath an undefined command
         c2.opt<int>("b", 99).command("three");
 
+        EXPECT_PARSE(c1, "one -a3");
+        EXPECT(*a1 == 3);
+        EXPECT(*a2 == 2);
+        EXPECT(c2.runCommand() == "one");
+        EXPECT_PARSE(c1, "-a", false);
+        EXPECT_ERR(c2, "Error: Unknown option: -a\n");
+        EXPECT_PARSE(c1, "two -a", false);
+        EXPECT_ERR(c2, "Error: Command 'two': Option requires value: -a\n");
+        EXPECT(c1.resetValues().exec() == false);
+        EXPECT_ERR(c1, "Error: No command given.\n");
+        EXPECT_PARSE(c1, "one");
+        EXPECT(c1.exec() == false);
+        EXPECT_ERR(c1, "Error: Command 'one' has not been implemented.\n");
+
         EXPECT_HELP(c1, "one", 1 + R"(
 usage: test one [OPTIONS]
 
@@ -502,14 +540,6 @@ Additional:
 Options:
   --help    Show this message and exit.
 )");
-        EXPECT_PARSE(c1, {"one", "-a3"});
-        EXPECT(*a1 == 3);
-        EXPECT(*a2 == 2);
-        EXPECT(c2.runCommand() == "one");
-        EXPECT_PARSE2(c1, false, Dim::kExitUsage, {"-a"});
-        EXPECT(c2.errMsg() == "Unknown option: -a");
-        EXPECT_PARSE2(c1, false, Dim::kExitUsage, {"two", "-a"});
-        EXPECT(c2.errMsg() == "Command 'two': Option requires value: -a");
     }
 
     // subcommand groups
@@ -564,22 +594,31 @@ Options:
   --help                    Show this message and exit.
 )";
         EXPECT_HELP(cli, "help", helpText);
-        EXPECT_PARSE2(cli, true, Dim::kExitOk, {"help", "help"});
+        EXPECT_PARSE(cli, "help help");
         out.str("");
         cli.iostreams(nullptr, &out);
         EXPECT(cli.exec());
-        EXPECT(out.str() == helpText);
         cli.iostreams(nullptr, nullptr);
+        EXPECT(out.str() == helpText);
         EXPECT_USAGE(cli, "", 1 + R"(
 usage: test [--help] command [args...]
 )");
         EXPECT_USAGE(cli, "help", 1 + R"(
 usage: test help [-u, --usage] [--help] [command]
 )");
-        EXPECT_PARSE(cli, {"help", "notACmd"});
+        EXPECT_PARSE(cli, "help notACmd");
         EXPECT(!cli.exec());
-        EXPECT(cli.errMsg() == 1 + R"(
-Command 'help': Help requested for unknown command: notACmd)");
+        EXPECT_ERR(cli, 1 + R"(
+Error: Command 'help': Help requested for unknown command: notACmd
+)");
+        EXPECT_PARSE(cli, "help help --usage");
+        out.str("");
+        cli.iostreams(nullptr, &out);
+        EXPECT(cli.exec());
+        cli.iostreams(nullptr, nullptr);
+        EXPECT(out.str() == 1 + R"(
+usage: test help [-u, --usage] [--help] [command]
+)");
     }
 }
 
@@ -623,19 +662,19 @@ void optCheckTests() {
     {
         cli = {};
         auto & count = cli.opt<int>("c", 1).require();
-        EXPECT_PARSE(cli, {"-c10"});
+        EXPECT_PARSE(cli, "-c10");
         EXPECT(*count == 10);
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {});
+        EXPECT_PARSE(cli, {}, false);
         EXPECT(*count == 1);
-        EXPECT(cli.errMsg() == "No value given for -c");
+        EXPECT_ERR(cli, "Error: No value given for -c\n");
         cli = {};
         auto & imp = cli.opt<int>("?index i").require().implicitValue(5);
-        EXPECT_PARSE(cli, {"--index=10"});
+        EXPECT_PARSE(cli, "--index=10");
         EXPECT(*imp == 10);
-        EXPECT_PARSE(cli, {"--index"});
+        EXPECT_PARSE(cli, "--index");
         EXPECT(*imp == 5);
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {});
-        EXPECT(cli.errMsg() == "No value given for --index");
+        EXPECT_PARSE(cli, {}, false);
+        EXPECT_ERR(cli, "Error: No value given for --index\n");
     }
 
     // clamp and range
@@ -643,15 +682,15 @@ void optCheckTests() {
         cli = {};
         auto & count = cli.opt<int>("<count>", 2).clamp(1, 10);
         auto & letter = cli.opt<char>("<letter>").range('a', 'z');
-        EXPECT_PARSE(cli, {"20", "a"});
+        EXPECT_PARSE(cli, "20 a");
         EXPECT(*count == 10);
         EXPECT(*letter == 'a');
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"5", "0"});
+        EXPECT_PARSE(cli, "5 0", false);
         EXPECT(*count == 5);
-        EXPECT(cli.errMsg() == "Out of range 'letter' value [a - z]: 0");
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"--", "-5"});
+        EXPECT_ERR(cli, "Error: Out of range 'letter' value [a - z]: 0\n");
+        EXPECT_PARSE(cli, "-- -5", false);
         EXPECT(*count == 1);
-        EXPECT(cli.errMsg() == "Missing argument: letter");
+        EXPECT_ERR(cli, "Error: Missing argument: letter\n");
     }
 }
 
@@ -690,7 +729,7 @@ Type of fruit:
 Other:
   --help    Show this message and exit.
 )");
-        EXPECT_PARSE(cli, {"-o"});
+        EXPECT_PARSE(cli, "-o");
         EXPECT(*orange == "orange");
         EXPECT(orange.from() == "-o");
         EXPECT(orange.pos() == 1 && orange.size() == 1);
@@ -700,7 +739,7 @@ Other:
         cli = {};
         auto & on = cli.opt<bool>("on.", true).flagValue();
         auto & notOn = cli.opt(on, "!notOn.").flagValue(true);
-        EXPECT_PARSE(cli, {"--on"});
+        EXPECT_PARSE(cli, "--on");
         EXPECT(*notOn);
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS]
@@ -717,7 +756,7 @@ Options:
         cli = {};
         auto & opt = cli.opt<int>("x", 1).flagValue(true);
         cli.opt(opt, "y", 2).flagValue();
-        EXPECT_PARSE(cli, {"-y"});
+        EXPECT_PARSE(cli, "-y");
         EXPECT(*opt == 2);
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS]
@@ -728,7 +767,7 @@ Options:
 
   --help    Show this message and exit.
 )");
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"-w"});
+        EXPECT_PARSE(cli, "-w", false);
     }
 
     {
@@ -736,7 +775,7 @@ Options:
         vector<int> vals;
         cli.optVec(&vals, "x").defaultValue(1).flagValue();
         cli.optVec(&vals, "y").defaultValue(2).flagValue(true);
-        EXPECT_PARSE(cli, {"-x"});
+        EXPECT_PARSE(cli, "-x");
         EXPECT(vals == vector<int>{1});
     }
 }
@@ -778,38 +817,38 @@ void responseTests() {
 
     cli = {};
     auto & args = cli.optVec<string>("[args]");
-    EXPECT_PARSE(cli, {"@test/a.rsp"});
+    EXPECT_PARSE(cli, "@test/a.rsp");
     EXPECT(*args == vector<string>{"1", "x", "y", "2"});
 
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"@test/does_not_exist.rsp"});
-    EXPECT(cli.errMsg() == "Invalid response file: test/does_not_exist.rsp");
+    EXPECT_PARSE(cli, "@test/does_not_exist.rsp", false);
+    EXPECT_ERR(cli, "Error: Invalid response file: test/does_not_exist.rsp\n");
     cli.responseFiles(false);
-    EXPECT_PARSE(cli, {"@test/does_not_exist.rsp"});
+    EXPECT_PARSE(cli, "@test/does_not_exist.rsp");
     EXPECT(args && args[0] == "@test/does_not_exist.rsp");
     cli.responseFiles(true);
 
 #ifdef _MSC_VER
-    EXPECT_PARSE(cli, {"@test/cL.rsp", "@test/du.rsp", "@test/f.rsp"});
+    EXPECT_PARSE(cli, "@test/cL.rsp @test/du.rsp @test/f.rsp");
     EXPECT(*args == vector<string>{"c1", "c2", "d1", "d2", "f"});
 #else
-    EXPECT_PARSE(cli, {"@test/f.rsp"});
+    EXPECT_PARSE(cli, "@test/f.rsp");
     EXPECT(*args == vector<string>{"f"});
 #endif
 
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"@test/gBad.rsp"});
-    EXPECT(cli.errMsg() == "Invalid encoding: eBad.rsp");
+    EXPECT_PARSE(cli, "@test/gBad.rsp", false);
+    EXPECT_ERR(cli, "Error: Invalid encoding: eBad.rsp\n");
 
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"@test/reA.rsp"});
-    EXPECT(cli.errMsg() == "Recursive response file: reA.rsp");
+    EXPECT_PARSE(cli, "@test/reA.rsp", false);
+    EXPECT_ERR(cli, "Error: Recursive response file: reA.rsp\n");
 
-    EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"@test/reX.rsp"});
-    EXPECT(cli.errMsg() == "Recursive response file: reX.rsp");
+    EXPECT_PARSE(cli, "@test/reX.rsp", false);
+    EXPECT_ERR(cli, "Error: Recursive response file: reX.rsp\n");
 
 #ifdef _MSC_VER
     {
         fstream f("test/f.rsp", ios::in, _SH_DENYRW);
-        EXPECT_PARSE2(cli, false, Dim::kExitUsage, {"@test/f.rsp"});
-        EXPECT(cli.errMsg() == "Read error: test/f.rsp");
+        EXPECT_PARSE(cli, "@test/f.rsp", false);
+        EXPECT_ERR(cli, "Error: Read error: test/f.rsp\n");
     }
 #endif
 
@@ -834,6 +873,15 @@ void basicTests() {
     {
         Dim::Cli tmp{cli};
         tmp = cli;
+    }
+
+    {
+        cli = {};
+        cli.helpOpt().show(false);
+        EXPECT_PARSE(cli);
+        EXPECT_HELP(cli, "", 1 + R"(
+usage: test
+)");
     }
 
     {
@@ -877,32 +925,32 @@ Name options:
 usage: test [-c COUNT] [-n, --number=NUM] [--n2=NUM] [--n3=NUM] [--name=STRING]
             [-s, --special] [--help] [key...]
 )");
-        EXPECT_PARSE(cli, {"-n3"});
+        EXPECT_PARSE(cli, "-n3");
         EXPECT(*num == 3);
         EXPECT(!*special);
         EXPECT(!name);
         EXPECT(!keys);
 
-        EXPECT_PARSE(cli, {"--name", "two"});
+        EXPECT_PARSE(cli, "--name two");
         EXPECT(*num == 0);
         EXPECT(name.size() == 1 && (*name)[0] == "two");
 
-        EXPECT_PARSE(cli, {"--name=three"});
+        EXPECT_PARSE(cli, "--name=three");
         EXPECT(name.size() == 1 && (*name)[0] == "three");
 
-        EXPECT_PARSE(cli, {"--name=", "key"});
+        EXPECT_PARSE(cli, "--name= key");
         EXPECT(*name == vector<string>({""s}));
         EXPECT(*keys == vector<string>({"key"s}));
 
-        EXPECT_PARSE(cli, {"-s-name=four", "key", "--name", "four"});
+        EXPECT_PARSE(cli, "-s-name=four key --name four");
         EXPECT(*special);
         EXPECT(*name == vector<string>({"four"s, "four"s}));
         EXPECT(*keys == vector<string>({"key"s}));
 
-        EXPECT_PARSE(cli, {"key", "extra"});
+        EXPECT_PARSE(cli, "key extra");
         EXPECT(*keys == vector<string>({"key"s, "extra"s}));
 
-        EXPECT_PARSE(cli, {"-", "--", "-s"});
+        EXPECT_PARSE(cli, "- -- -s");
         EXPECT(!special && !*special);
         *num += 2;
         EXPECT(*num == 2);
@@ -912,12 +960,20 @@ usage: test [-c COUNT] [-n, --number=NUM] [--n2=NUM] [--n3=NUM] [--name=STRING]
 
     {
         cli = {};
+        cli.opt<int>("[]", 1);
+        EXPECT_USAGE(cli, "", 1 + R"(
+usage: test [--help] [arg1]
+)");
+    }
+
+    {
+        cli = {};
         cli.opt<bool>("option with excessively long list of key names")
             .desc("Why so many?");
-        cli.opt<string>("[equal '=' in name]")
+        cli.opt<string>("< equal '=' in name >")
             .desc("Equals are sometimes reserved.");
         EXPECT_HELP(cli, "", 1 + R"(
-usage: test [OPTIONS] [<equal '=' in name>]
+usage: test [OPTIONS] <equal '=' in name>
   equal '=' in name  Equals are sometimes reserved.
 
 Options:
@@ -934,7 +990,7 @@ Options:
         cli = {};
         auto & strs = cli.optVec<string>("r ?s").implicitValue("a")
             .desc("String array.");
-        EXPECT_PARSE(cli, {"-s1", "-s", "-r", "2", "-s3"});
+        EXPECT_PARSE(cli, "-s1 -s -r 2 -s3");
         EXPECT(strs.size() == 4);
         EXPECT(strs.pos(2) == 4);
         EXPECT(strs.pos() == 5);
@@ -958,7 +1014,7 @@ Options:
 
   --help           Show this message and exit.
 )");
-        EXPECT_PARSE(cli, {"--string=a", "-sb"});
+        EXPECT_PARSE(cli, "--string=a -sb");
         EXPECT(*strs == vector<string>({"a"s, "b"s}));
     }
 
@@ -972,7 +1028,7 @@ Options:
         os << path;
         cli.opt(&path, "path", path)
             .desc("std::filesystem::path");
-        EXPECT_PARSE(cli, {"--path", "one"});
+        EXPECT_PARSE(cli, "--path one");
         EXPECT(path == "one");
         EXPECT_HELP(cli, "", 1 + R"(
 usage: test [OPTIONS]
@@ -1007,21 +1063,27 @@ Options:
 
   --help             Show this message and exit.
 )");
-        EXPECT_PARSE(cli, {"--password=hi"});
+        EXPECT_PARSE(cli, "--password=hi");
         EXPECT(*pass == "hi");
         in.clear();
         in.str("secret\nsecret\n");
         out.str("");
         cli.iostreams(&in, &out);
-        EXPECT_PARSE(cli, {});
+        EXPECT_PARSE(cli);
         EXPECT(*pass == "secret");
         EXPECT(out.str() == "Password: \nEnter again to confirm: \n");
+        EXPECT(in.get() == EOF);
+        in.clear();
+        in.str("secret\nmistype_secret\n");
+        out.str("");
+        EXPECT_PARSE(cli, {}, false);
+        EXPECT_ERR(cli, "Error: Confirm failed, entries not the same.\n");
         EXPECT(in.get() == EOF);
         if (prompt) {
             cli.iostreams(nullptr, nullptr);
             cout << "Expects password to be confirmed (empty is ok)." << endl
                  << "What you type should *NOT* be visible!" << endl;
-            EXPECT_PARSE(cli, {});
+            EXPECT_PARSE(cli);
             cout << "Entered password was '" << *pass << "'" << endl;
         }
     }
@@ -1030,13 +1092,13 @@ Options:
     {
         cli = {};
         auto & ask = cli.confirmOpt();
-        EXPECT_PARSE(cli, {"-y"});
+        EXPECT_PARSE(cli, "-y");
         EXPECT(*ask);
         in.clear();
         in.str("n\n");
         out.str("");
         cli.iostreams(&in, &out);
-        EXPECT_PARSE2(cli, false, Dim::kExitOk, {});
+        EXPECT_PARSE(cli, {}, false, Dim::kExitOk);
         EXPECT(!*ask);
         EXPECT(out.str() == "Are you sure? [y/N]: ");
         EXPECT(in.get() == EOF);
@@ -1044,7 +1106,7 @@ Options:
             cli.iostreams(nullptr, nullptr);
             cout << "Expects answer to be no." << endl
                  << "What you type should be visible." << endl;
-            EXPECT_PARSE2(cli, false, Dim::kExitOk, {});
+            EXPECT_PARSE(cli, {}, false, Dim::kExitOk);
             EXPECT(!*ask);
         }
     }
@@ -1052,13 +1114,13 @@ Options:
     // prompt for array
     {
         cli = {};
-        auto & ask = cli.optVec<string>("name").prompt();
+        auto & ask = cli.optVec<string>("name").prompt(Dim::Cli::fPromptHide);
         in.clear();
         in.str("jack\n");
         out.str("");
         cli.iostreams(&in, &out);
-        EXPECT_PARSE(cli, {});
-        EXPECT(out.str() == "Name: ");
+        EXPECT_PARSE(cli);
+        EXPECT(out.str() == "Name: \n");
         EXPECT(ask.size() && ask[0] == "jack");
     }
 
@@ -1070,7 +1132,7 @@ Options:
         in.str("jack\n");
         out.str("");
         cli.iostreams(&in, &out);
-        EXPECT_PARSE(cli, {});
+        EXPECT_PARSE(cli);
         EXPECT(out.str() == "Name [jill]: ");
         EXPECT(*ask == "jack");
     }
@@ -1088,12 +1150,12 @@ void envTests() {
     cli.envOpts("TEST_OPTS");
     result = putenv((char *) "TEST_OPTS=");
     EXPECT(!result);
-    EXPECT_PARSE(cli, {"c", "d"});
-    EXPECT(args->size() == 2);
+    EXPECT_PARSE(cli, "c d");
+    EXPECT(*args == vector<string>{"c", "d"});
     result = putenv((char *) "TEST_OPTS=a b");
     EXPECT(!result);
-    EXPECT_PARSE(cli, {"c", "d"});
-    EXPECT(args->size() == 4);
+    EXPECT_PARSE(cli, "c d");
+    EXPECT(*args == vector<string>{"a", "b", "c", "d"});
 #endif
 }